Move IpAddress reducer
In the process create tests for the ipAddress
reducer

close DCOS-14695
