+++ plugins/services/src/js/reducers/JSONMultiContainerReducers.js
@@ -6,7 +6,7 @@ import {JSONReducer as scaling} from './serviceForm/MultiContainerScaling';
 import {JSONReducer as labels} from './serviceForm/Labels';
 import {JSONReducer as volumes} from './serviceForm/MultiContainerVolumes';
 import {JSONReducer as networks} from './serviceForm/MultiContainerNetwork';
-import {JSONReducer as ipAddress} from './serviceForm/IpAddress';
+import {JSONReducer as ipAddress} from './serviceForm/JSONReducers/IpAddress';
 import {
   simpleReducer
 } from '../../../../../src/js/utils/ReducerUtil';
