+++ plugins/services/src/js/reducers/JSONSingleContainerReducers.js
@@ -6,7 +6,7 @@ import {JSONReducer as healthChecks} from './serviceForm/HealthChecks';
 import {JSONReducer as labels} from './serviceForm/Labels';
 import {JSONReducer as portDefinitions} from './serviceForm/PortDefinitions';
 import {JSONReducer as residency} from './serviceForm/Residency';
-import {JSONReducer as ipAddress} from './serviceForm/IpAddress';
+import {JSONReducer as ipAddress} from './serviceForm/JSONReducers/IpAddress';
 import {JSONReducer as requirePorts} from './serviceForm/RequirePorts';
 import {
   simpleFloatReducer,
