+++ plugins/services/src/js/reducers/serviceForm/JSONReducers/IpAddress.js
@@ -1,4 +1,4 @@
-import {SET} from '../../../../../../src/js/constants/TransactionTypes';
+import {SET} from '../../../../../../../src/js/constants/TransactionTypes';
 
 module.exports = {
   JSONReducer(state = null, {type, path, value}) {
